Move JSON serialization code to Json folder
