+++ src/Razor/src/Microsoft.AspNetCore.Razor.ProjectEngineHost/Serialization/Json/Converters/ProjectSnapshotHandleJsonConverter.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
-namespace Microsoft.AspNetCore.Razor.Serialization.Converters;
+namespace Microsoft.AspNetCore.Razor.Serialization.Json.Converters;
 
 internal partial class ProjectSnapshotHandleJsonConverter : ObjectJsonConverter<ProjectSnapshotHandle>
 {
