+++ src/Razor/test/Microsoft.AspNetCore.Razor.ProjectEngineHost.Test/Serialization/TagHelperDeltaResultSerializationTest.cs
@@ -9,7 +9,8 @@ using System.Text;
 using Microsoft.AspNetCore.Mvc.Razor.Extensions;
 using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.AspNetCore.Razor.Serialization;
-using Microsoft.AspNetCore.Razor.Serialization.Converters;
+using Microsoft.AspNetCore.Razor.Serialization.Json;
+using Microsoft.AspNetCore.Razor.Serialization.Json.Converters;
 using Microsoft.AspNetCore.Razor.Test.Common;
 using Newtonsoft.Json;
 using Xunit;
